wip: porting settings modules into separate crate
