+++ Cargo.toml
@@ -2,6 +2,7 @@
 members = [
     "cli",
     "dbus",
+    "settings",
     "services",
     "nats",
 ]
