+++ services/Cargo.toml
@@ -24,7 +24,6 @@ console = "0.14"
 dialoguer = "0.8"
 file-lock = "2.1.4"
 futures = "0.3"
-git2 = "0.15.0"                 # Bindings to libgit2 for interoperating with git repositories.
 glob = "0.3.0"
 hex = "0.4"
 http = "0.2.5"
@@ -34,11 +33,11 @@ log = "0.4"
 parking_lot = "0.12.1"                  # More compact and efficient implementations of the standard synchronization primitives.
 printnanny-api-client = "0.113.0"
 printnanny-dbus = { path = "../dbus", version = "^0.1" }
+printnanny-settings = { path = "../settings", version = "^0.1"}
 procfs = "0.12"
 rustls = "0.19"
 tokio-rustls = "0.22"
 sha2 = "0.9.8"
-sys-info = "0.9"
 rand = "0.8"
 reqwest = "0.11"
 sysinfo = "0.26.2"
