+++ settings/Cargo.toml
@@ -2,17 +2,30 @@
 name = "printnanny-settings"
 version = "0.1.0"
 edition = "2021"
+authors = ["Leigh Johnson <leigh@printnanny.ai>"]
+license = "AGPL-3.0-or-later"
+description = "Utilities for interacting with PrintNanny settings"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+rust-version = "1.63"
 
 [dependencies]
+async-trait = "0.1"
+printnanny-asyncapi-models = { version = "0.1.15"    } # PritnNanny AsyncAPI Rust models generated by Modelina
 serde = { version = "1", features = ["derive"] }
 clap = { version = "3", features = ["derive", "cargo", "env", "wrap_help"] }
 figment = { version = "0.10", features = ["env", "json", "toml", "yaml"] }
-serde = { version = "1", features = ["derive"] }
+log = "0.4"
 toml = "0.5"
+printnanny-dbus = { path = "../dbus", version = "^0.1" }
+serde_json = "1"
+sys-info = "0.9"
+thiserror = "1"
+
+git2 = "0.15.0"                 # Bindings to libgit2 for interoperating with git repositories.
 
 [dev-dependencies]
 predicates = "2"
 figment = { version = "0.10", features = ["env", "json", "toml", "test"] }
 test-log = "0.2"
+env_logger = "0.9"
