+++ settings/src/moonraker.rs
@@ -10,8 +10,9 @@ use serde::{Deserialize, Serialize};
 use printnanny_dbus::zbus;
 use printnanny_dbus::zbus_systemd;
 
-use super::vcs::{VersionControlledSettings, VersionControlledSettingsError};
-use crate::settings::SettingsFormat;
+use crate::error::VersionControlledSettingsError;
+use crate::vcs::VersionControlledSettings;
+use crate::SettingsFormat;
 
 pub const MOONRAKER_INSTALL_DIR: &str = "/home/printnanny/.moonraker";
 pub const MOONRAKER_VENV: &str = "/home/printnanny/moonraker-venv";
