+++ settings/src/vcs.rs
@@ -6,34 +6,10 @@ use git2::{DiffFormat, DiffOptions, Repository};
 use log::info;
 use printnanny_asyncapi_models::SettingsFile;
 use serde::{Deserialize, Serialize};
-use thiserror::Error;
 
-use printnanny_dbus::zbus;
-
-use crate::error::PrintNannyCloudDataError;
-use crate::error::PrintNannySettingsError;
-use crate::settings::printnanny::PrintNannySettings;
-use crate::settings::SettingsFormat;
-
-#[derive(Error, Debug)]
-pub enum VersionControlledSettingsError {
-    #[error("Failed to write {path} - {error}")]
-    WriteIOError { path: String, error: std::io::Error },
-    #[error("Failed to read {path} - {error}")]
-    ReadIOError { path: String, error: std::io::Error },
-    #[error("Failed to copy {src:?} to {dest:?} - {error}")]
-    CopyIOError {
-        src: PathBuf,
-        dest: PathBuf,
-        error: std::io::Error,
-    },
-    #[error(transparent)]
-    GitError(#[from] git2::Error),
-    #[error(transparent)]
-    ZbusError(#[from] zbus::Error),
-    #[error(transparent)]
-    PrintNannyCloudDataError(#[from] PrintNannyCloudDataError),
-}
+use crate::error::VersionControlledSettingsError;
+use crate::printnanny::PrintNannySettings;
+use crate::SettingsFormat;
 
 #[derive(Clone, Debug, Eq, PartialEq, Serialize, Deserialize)]
 pub struct GitCommit {
